fix inconsistencys all project

+++ database/migrations/2026_03_10_125139_create_transactions_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('gateway_id')->constrained()->onDelete('cascade');
             $table->string('external_id')->nullable();
             $table->enum('status', ['pending', 'completed', 'failed','refunded'])->default('pending');
-            $table->integer('amount');
+            $table->decimal('amount', 10, 2);
             $table->string('quantity')->nullable();
             $table->enum('payment_method', ['card_credit', 'card_debit', 'pix', 'boleto']);
             $table->char('cvv', 4);
